% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{ACC_BUSINESS_TYPE}
\alias{ACC_BUSINESS_TYPE}
\title{ACCESSIBILITY_DATA
ACC_BUSINESS_TYPE}
\format{
An object of class \code{character} of length 14.
}
\usage{
ACC_BUSINESS_TYPE
}
\description{
ACCESSIBILITY_DATA
ACC_BUSINESS_TYPE
}
\keyword{datasets}
