% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{ACC_PARAM_LIST}
\alias{ACC_PARAM_LIST}
\title{ACC_COLOR_LEVEL_LIST}
\format{
An object of class \code{list} of length 5.
}
\usage{
ACC_PARAM_LIST
}
\description{
ACC_COLOR_LEVEL_LIST
}
\keyword{datasets}
