% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{ACC_TRANSPORTATION_METHODS}
\alias{ACC_TRANSPORTATION_METHODS}
\title{ACC_TRANSPORTATION_METHODS}
\format{
An object of class \code{character} of length 3.
}
\usage{
ACC_TRANSPORTATION_METHODS
}
\description{
ACC_TRANSPORTATION_METHODS
}
\keyword{datasets}
