% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{BIRTH_OUTCOMES_PAL}
\alias{BIRTH_OUTCOMES_PAL}
\title{BIRTH_OUTCOMES_PAL}
\format{
An object of class \code{character} of length 7.
}
\usage{
BIRTH_OUTCOMES_PAL
}
\description{
BIRTH_OUTCOMES_PAL
}
\keyword{datasets}
