% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_files_update_time}
\alias{check_files_update_time}
\title{Check the updated time of files in a specified folder}
\usage{
check_files_update_time(dir, regexp)
}
\arguments{
\item{dir}{file directory}

\item{regexp}{A regular expression to filter paths}
}
\value{
return the a tibble with the first column \code{path} contains all of the
matched files path and second column \code{modification_time} contains the modification time of the file
}
\description{
Check the updated time of files in a specified folder
}
