% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{DATA_SOURCE_DOC_Mahoning}
\alias{DATA_SOURCE_DOC_Mahoning}
\title{Data source documentation Mahoning}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 4 columns.
}
\usage{
DATA_SOURCE_DOC_Mahoning
}
\description{
Data source documentation Mahoning
}
\keyword{datasets}
