% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{DISEASE_DATA}
\alias{DISEASE_DATA}
\title{Processed disease data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 70 rows and 39 columns.
}
\usage{
DISEASE_DATA
}
\description{
Processed disease data
}
\keyword{datasets}
