% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{DISEASE_DATA_VARS}
\alias{DISEASE_DATA_VARS}
\title{Disease data vars}
\format{
An object of class \code{character} of length 12.
}
\usage{
DISEASE_DATA_VARS
}
\description{
Disease data vars
}
\keyword{datasets}
