% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{DISEASE_OUTCOMES}
\alias{DISEASE_OUTCOMES}
\title{Disease data outcomes}
\format{
An object of class \code{character} of length 12.
}
\usage{
DISEASE_OUTCOMES
}
\description{
Disease data outcomes
}
\keyword{datasets}
