% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{EQUITY_MAP_FILTER_CHOICES}
\alias{EQUITY_MAP_FILTER_CHOICES}
\title{EQUITY_MAP_FILTER_CHOICES}
\format{
An object of class \code{list} of length 10.
}
\usage{
EQUITY_MAP_FILTER_CHOICES
}
\description{
EQUITY_MAP_FILTER_CHOICES
}
\keyword{datasets}
