% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{FILTER_CITY_CHOICES}
\alias{FILTER_CITY_CHOICES}
\title{City Choices for business location filter}
\format{
An object of class \code{character} of length 48.
}
\usage{
FILTER_CITY_CHOICES
}
\description{
City Choices for business location filter
}
\keyword{datasets}
