% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{FILTER_TYPE_CHOICES}
\alias{FILTER_TYPE_CHOICES}
\title{Business type Choices for business location filter}
\format{
An object of class \code{character} of length 41.
}
\usage{
FILTER_TYPE_CHOICES
}
\description{
Business type Choices for business location filter
}
\keyword{datasets}
