% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{FILTER_ZIP_CHOICES}
\alias{FILTER_ZIP_CHOICES}
\title{Zip Choices for business location filter}
\format{
An object of class \code{character} of length 63.
}
\usage{
FILTER_ZIP_CHOICES
}
\description{
Zip Choices for business location filter
}
\keyword{datasets}
