% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_map_shapefile}
\alias{load_map_shapefile}
\title{Load Map Shapefile Function}
\usage{
load_map_shapefile(file, crs = 4326)
}
\arguments{
\item{file}{sf file path}

\item{crs}{crs code default to be 4326}
}
\description{
Load Map Shapefile Function
}
