% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{PAL}
\alias{PAL}
\title{Store all tokens used in the app
PAL}
\format{
An object of class \code{list} of length 10.
}
\usage{
PAL
}
\description{
Store all tokens used in the app
PAL
}
\keyword{datasets}
