% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{ROOT}
\alias{ROOT}
\title{Root Dir Path}
\format{
An object of class \code{fs_path} (inherits from \code{character}) of length 1.
}
\usage{
ROOT
}
\description{
Root Dir Path
}
\keyword{datasets}
