% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App-server.R
\name{server}
\alias{server}
\title{Shiny server}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny outptu}

\item{session}{shiny session}
}
\description{
Shiny server
}
