% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{SF_CENSUS_TRACT}
\alias{SF_CENSUS_TRACT}
\title{Census tract shapefile for Mahoning County}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 70 rows and 16 columns.
}
\usage{
SF_CENSUS_TRACT
}
\description{
Census tract shapefile for Mahoning County
}
\keyword{datasets}
