% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{SF_ZIP}
\alias{SF_ZIP}
\title{ZIP shapefile for Mahoning County}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 43 rows and 11 columns.
}
\usage{
SF_ZIP
}
\description{
ZIP shapefile for Mahoning County
}
\keyword{datasets}
