% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{TOKENS}
\alias{TOKENS}
\title{Store all tokens used in the app}
\format{
An object of class \code{list} of length 3.
}
\usage{
TOKENS
}
\description{
Store all tokens used in the app
}
\keyword{datasets}
