% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{ACCESSIBILITY_DATA}
\alias{ACCESSIBILITY_DATA}
\title{ACCESSIBILITY_DATA}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 980 rows and 7 columns.
}
\usage{
ACCESSIBILITY_DATA
}
\description{
ACCESSIBILITY_DATA
}
\keyword{datasets}
