% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc.R
\docType{data}
\name{BUSINESS_LOCATION_DATA}
\alias{BUSINESS_LOCATION_DATA}
\title{BUSINESS LOCATION DATA}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1370 rows and 13 columns.
}
\usage{
BUSINESS_LOCATION_DATA
}
\description{
BUSINESS LOCATION DATA
}
\keyword{datasets}
